requirements.txt
roboflow==1.0.9
python-dotenv==1.0.0
ultralytics==8.0.20
opencv-python==4.7.0.72